+++ src/App.tsx
@@ -2,13 +2,13 @@ import 'bootstrap/dist/css/bootstrap.css'
 import { createBrowserRouter, RouterProvider } from 'react-router-dom'
 import './App.css'
 import TemplateEditor from './components/editor'
-import Home from './components/home'
+import WorkFlowView from './components/workflow_view'
 import JobsView from './components/jobs_view'
 import JsonStoreView from './components/json_store_view'
 import MdxTemplateView from './components/mdx_template_view'
 import RootLayout from './components/root_layout'
 import templatesConfig from './templates/templates_config'
-import ProjectsListView from './components/projects_list_view'
+import Home from './components/home'
 import Projects from './templates/projects_config'
 import MdxProjectView from './components/mdx_project_details_view'
 
@@ -20,7 +20,7 @@ const routes = [
         path: '/',
         element: (
             <RootLayout>
-                <ProjectsListView />
+                <Home />
             </RootLayout>
         ),
     },
@@ -36,7 +36,7 @@ const routes = [
                 path: `/app/${key?._id}/workflows`,
                 element: (
                     <RootLayout>
-                        <Home project={key} />
+                        <WorkFlowView project={key} />
                     </RootLayout>
                 ),
             },
