+++ src/components/string_input_modal.tsx
@@ -43,6 +43,7 @@ const StringInputModal: React.FC<StringInputModalProps> = ({
 
     const handleSubmit = () => {
         onSubmit(inputValue.trim())
+        refreshPage()
         closeModal()
     }
 
@@ -75,6 +76,10 @@ const StringInputModal: React.FC<StringInputModalProps> = ({
             const errorMessage = errorValidator?.errorMsg || ''
             setErrorMessage(errorMessage)
         }
+    }
+
+    function refreshPage() {
+        window.location.reload();
     }
 
     const modalTitle = title || 'Default Title'
