+++ src/templates/doe_workflow_hpwh.mdx
@@ -202,9 +202,7 @@
 
     <ProjectInfoReport {...props} />
    
-    <Photo id="building_number_photo" label="Building Number" required>
-      A photo of the building showing the building number.
-    </Photo>
+   
 
     ## About the Report
     
@@ -264,18 +262,14 @@
     }
        
 
-    {
-      ( props.metadata?.attachments?.old_nameplate_photo != null )
-      ? <>
-          <h2>Old Hot Water Heater</h2>
-          <Photo id="old_nameplate_photo" label="3. Old Water Heater Nameplate – Photo">
-            This photo shows the name plate of the old water heater. It may be used for rebate purposes
-            or to determine the energy savings.
-          </Photo>
-        </>
-      : null 
-    }
-
+  
+    <h2>Old Hot Water Heater</h2>
+    <Photo id="old_nameplate_photo" label="3. Old Water Heater Nameplate – Photo">
+      This photo shows the name plate of the old water heater. It may be used for rebate purposes
+      or to determine the energy savings.
+    </Photo>
+        
+   
     ## Installation Day Photos
     <Photo id="completed_installed_photo" label="4. Completed Installation – Photo" required>
       This wide-angle photo shows the completed HPWH installation including the exhaust,
