+++ src/templates/reusable/project_info_inputs.mdx
@@ -20,3 +20,5 @@
     <StringInput label="Company Address" path="installer.mailing_address" />
     <StringInput label="Company Phone" path="installer.phone" />
     <StringInput label="Company Email" path="installer.email" />
+
+
