+++ src/templates/reusable/project_info_report.mdx
@@ -10,3 +10,7 @@
 **Installation Date:**  <DateStr date={props.data.install_date}/> 
 
 **Installation Address:** {props.project.data_.location?.street_address}, {props.project.data_.location?.city}, {props.project.data_.location?.state} {props.project.data_.location?.zip_code}
+
+<Photo id="building_number_photo" label="Building Number" required>
+      A photo of the building showing the building number.
+</Photo>
